Add an ast module.
The ast module mirrors the Python's builtin ast module, but adds
location information. Specifically, in the same way that the builtin
module describes the semantic fields, it adds a class field _locs
containing the name of the fields containing location data.

Additionally, this module contains documentation, describing
just what the semantic (and location) fields contain.

The ast module currently targets 2.6-2.7, with shim infrastructure
to be later extended up to 3.5.