Add parser module.
Currently this only implements 2.6 grammar, but in full.
There are possible location bugs and semantic action bugs--
a test suite will be necessary to eradicate these--
but probably no grammar bugs, as it is mostly a straightforward
translation of the official Python grammar.
The latter also ensures it will be easy to update it for
newer Python versions.