Insert newline into token stream before EOF.
The Python grammar has three entry points:

    single_input: NEWLINE | simple_stmt | compound_stmt NEWLINE
    file_input: (NEWLINE | stmt)* ENDMARKER
    eval_input: testlist NEWLINE* ENDMARKER
    [...]
    stmt: simple_stmt | compound_stmt
    simple_stmt: small_stmt (';' small_stmt)* [';'] NEWLINE

A trailing newline is required in the input in every entry point except
eval_input. However, this is not how the real Python works. ast.parse() and
exec() both do not require a newline:

    >>> ast.parse('print "hola"')
    <_ast.Module object at 0x10425df50>
    >>> exec 'print "hola"'
    hola

CPython hacks this in by replacing the ENDMARKER with a NEWLINE in the token
stream: 
https://github.com/python/cpython/blob/af88e7eda4101f36e904771d3cf59a5f740b3b00/Parser/parsetok.c#L213

After this commit, we now perform a similar operation.